DateCustomTypeAdapterUnitTest, Move MockClientCallback
